requirements.txt
Django==1.8.3
djangorestframework==3.1.3
docker==3.2.1
